sulidashboard/requirements.txt
streamlit
plotly
pandas
pyvis
gspread
oauth2client
ipython
wordcloud
matplotlib
openpyxl
nltk
gspread
google-auth
google-auth-oauthlib
streamlit_autorefresh
